+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Settings validator to prevent attributes that inherit from BaseSettings
+
 ### Changes
 
 ### Deprecated